requirements.txt
asgiref==3.3.1
Django==3.1.7
django-appconf==1.0.4
django-bootstrap-modal-forms==2.1.0
django-select2==7.6.2
pytz==2021.1
sqlparse==0.4.1
